requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
boto3==1.40.24
botocore==1.40.24
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
frozenlist==1.7.0
idna==3.10
inflection==0.5.1
isort==6.0.1
jmespath==1.0.1
kombu==5.5.4
multidict==6.6.4
packaging==25.0
phonenumbers==9.0.13
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
twilio==9.8.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
